requirements.txt
Flask==3.1.0
langchain==0.3.23
langchain_community==0.3.21
langchain_pinecone==0.2.5
pinecone==6.0.2
python-dotenv==1.1.0
setuptools==65.6.3
setuptools==65.5.0
